Log min level must overwrite
